samples/custom-voice/python/requirements.txt
requests
azure-cognitiveservices-speech
python-dotenv
urllib3>=2.6.3
